Algorithm is now explained in docstring, docstring now linked to sphinx. Files moved to the physics module. Minor fixes in source code.
