+++ doc/src/modules/physics/hep/gamma_matrices.rst
@@ -1,0 +1,6 @@
+High energy physics
+===================
+
+.. module:: sympy.physics.hep.gamma_matrices
+
+.. autofunction:: sympy.physics.hep.gamma_matrices.kahane_simplify
