+++ doc/src/modules/physics/hep/index.rst
@@ -1,0 +1,15 @@
+===================
+High energy physics
+===================
+
+.. topic:: Abstract
+
+   Contains docstrings for methods in high energy physics.
+
+Gamma matrices
+==============
+
+.. toctree::
+    :maxdepth: 3
+
+    gamma_matrices.rst
